requirements.txt
fastapi == 0.63.0
Keras == 2.4.3
matplotlib == 3.2.2
numpy == 1.18.5
scikit-learn == 0.23.1
scipy == 1.5.0
h5py == 2.10.0
tensorflow-gpu == 2.3.1
Pillow == 7.2.0
pydantic == 1.8.1
typing-extensions == 3.7.4.3
uvicorn == 0.13.4
opencv-python